@@ -1,31 +1,69 @@
 {
   "name": "api-gateway",
   "version": "1.0.0",
-  "description": "API Gateway for microservices architecture",
-  "main": "src/index.js",
+  "description": "Modern API Gateway with TypeScript for microservices architecture",
+  "main": "dist/index.js",
   "type": "module",
   "scripts": {
-    "start": "node src/index.js",
-    "dev": "nodemon src/index.js",
-    "test": "jest"
+    "build": "tsc",
+    "start": "node dist/index.js",
+    "dev": "nodemon --exec ts-node src/index.ts",
+    "test": "jest",
+    "lint": "eslint . --ext .ts",
+    "format": "prettier --write \"src/**/*.ts\"",
+    "typecheck": "tsc --noEmit"
   },
   "dependencies": {
+    "@opentelemetry/api": "^1.7.0",
+    "@opentelemetry/auto-instrumentations-node": "^0.41.1",
+    "@opentelemetry/sdk-node": "^0.48.0",
+    "@types/express-validator": "^2.20.33",
+    "@types/jsonwebtoken": "^9.0.8",
+    "@types/redis": "^4.0.10",
+    "axios": "^1.7.9",
+    "compression": "^1.7.4",
     "cors": "^2.8.5",
     "express": "^4.18.2",
     "express-rate-limit": "^7.1.5",
-    "helmet": "^7.1.0",
-    "http-proxy-middleware": "^2.0.6",
+    "express-validator": "^7.2.1",
+    "helmet": "^7.2.0",
+    "http-proxy-middleware": "^2.0.7",
+    "http-status-codes": "^2.3.0",
+    "ioredis": "^5.3.2",
+    "jose": "^5.2.2",
     "jsonwebtoken": "^9.0.2",
+    "opentelemetry": "^0.1.0",
     "opossum": "^8.1.3",
+    "pino": "^8.18.0",
+    "pino-http": "^9.0.0",
+    "prom-client": "^15.1.0",
+    "rate-limit-redis": "^4.2.0",
+    "rate-limiter-flexible": "^5.0.5",
+    "redis": "^4.7.0",
     "winston": "^3.11.0",
-    "prom-client": "^14.2.0"
+    "zod": "^3.22.4"
   },
   "devDependencies": {
+    "@types/compression": "^1.7.5",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.11.19",
+    "@types/supertest": "^6.0.2",
+    "@typescript-eslint/eslint-plugin": "^7.0.1",
+    "@typescript-eslint/parser": "^7.0.1",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
     "jest": "^29.7.0",
     "nodemon": "^3.0.3",
-    "supertest": "^6.3.4"
+    "prettier": "^3.2.5",
+    "supertest": "^6.3.4",
+    "ts-jest": "^29.1.2",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3"
   },
   "engines": {
-    "node": ">=18.0.0"
+    "node": ">=20.0.0"
   }
 }
