@@ -1,7 +1,7 @@
 import { Request, Response, NextFunction } from 'express';
 import jwt from 'jsonwebtoken';
-import { AuthRequest, TokenPayload } from '../types';
-import logger from '../utils/logger';
+import { AuthRequest, TokenPayload } from '../types/index.js';
+import { logger } from '../utils/logger.js';
 
 export const authenticate = (
   req: AuthRequest,
@@ -9,84 +9,84 @@ export const authenticate = (
   next: NextFunction
 ) => {
   try {
-    // Authorization header'ını kontrol et
+    // Check Authorization header
     const authHeader = req.headers.authorization;
     if (!authHeader) {
       return res.status(401).json({
         error: true,
-        message: 'Yetkilendirme başlığı eksik'
+        message: 'Authorization header missing'
       });
     }
 
-    // Bearer token'ı ayıkla
+    // Extract Bearer token
     const token = authHeader.split(' ')[1];
     if (!token) {
       return res.status(401).json({
         error: true,
-        message: 'Token bulunamadı'
+        message: 'Token not found'
       });
     }
 
-    // Token'ı doğrula
+    // Verify token
     const decoded = jwt.verify(
       token,
       process.env.JWT_ACCESS_SECRET!
     ) as TokenPayload;
     
-    // Kullanıcı bilgisini request nesnesine ekle
+    // Add user info to request object
     req.user = decoded;
 
     next();
   } catch (error) {
     if (error instanceof jwt.TokenExpiredError) {
-      logger.warn('Süresi dolmuş token kullanım denemesi');
+      logger.warn('Expired token access attempt');
       return res.status(401).json({
         error: true,
-        message: 'Token süresi dolmuş'
+        message: 'Token expired'
       });
     }
 
     if (error instanceof jwt.JsonWebTokenError) {
-      logger.warn('Geçersiz token kullanım denemesi');
+      logger.warn('Invalid token access attempt');
       return res.status(401).json({
         error: true,
-        message: 'Geçersiz token'
+        message: 'Invalid token'
       });
     }
 
-    logger.error('Kimlik doğrulama hatası:', error);
+    logger.error('Authentication error:', error);
     res.status(500).json({
       error: true,
-      message: 'Kimlik doğrulama işlemi sırasında bir hata oluştu'
+      message: 'An error occurred during authentication'
     });
   }
 };
 
-// Belirli rollere sahip kullanıcıları kontrol et
+// Check for specific user roles
 export const authorize = (roles: string[] = []) => {
   return (req: AuthRequest, res: Response, next: NextFunction) => {
     try {
       if (!req.user) {
         return res.status(401).json({
           error: true,
-          message: 'Yetkilendirme gerekli'
+          message: 'Authorization required'
         });
       }
 
       if (roles.length && !roles.includes(req.user.role || '')) {
-        logger.warn(`Yetkisiz erişim denemesi - Kullanıcı: ${req.user.id}, İstenen rol: ${roles.join(', ')}`);
+        logger.warn(`Unauthorized access attempt - User: ${req.user.id}, Required roles: ${roles.join(', ')}`);
         return res.status(403).json({
           error: true,
-          message: 'Bu işlem için yetkiniz yok'
+          message: 'You do not have permission for this action'
         });
       }
 
       next();
     } catch (error) {
-      logger.error('Yetkilendirme hatası:', error);
+      logger.error('Authorization error:', error);
       res.status(500).json({
         error: true,
-        message: 'Yetkilendirme işlemi sırasında bir hata oluştu'
+        message: 'An error occurred during authorization'
       });
     }
   };
