@@ -1,6 +1,6 @@
 import { Sequelize } from 'sequelize';
 
-const sequelize = new Sequelize({
+export const sequelize = new Sequelize({
   dialect: 'postgres',
   host: process.env.DB_HOST || 'db',
   port: process.env.DB_PORT || 5432,
@@ -8,5 +8,3 @@ const sequelize = new Sequelize({
   username: process.env.DB_USER || 'postgres',
   password: process.env.DB_PASSWORD || 'postgres'
 });
-
-export default sequelize;
